requirements.txt
fastapi
uvicorn
pandas
openai
google-api-python-client
google-auth
python-multipart
